systemd support: Hook up installation in autotools setup.
Installs lirc as two separate systemd services lircd.socket
and lircmd.service. Systemd's requires/wants declarations
doesn't really fit the bill for an overall lirc service.

Part of this is a Fedora setup: the systemd environment is in
/etc/sysconfig/lirc, and this file is referenced in lircd.service.
This could be generalized with e. g., an user option to configure,
but as of now this is left to downstream to patch.