Adding configs directory.
The configs dir is the per driver configuration stuff which is
removed from configure.ac. As is, it should be fairly simple
to restore the script which generates the driver table. In the
logn run, this should be the base for a little more sophisticated
runtime setup up with default for lircd driver and ports,
including modprobe parameters support.