m4/ac-path-kernel-source.m4: Updates
This macro does not longer work. It does not have sensible
defaults for Debian, and the generated Makefiles does not
work due to kernel updates.

Try also debian defaults, and use simple grepping in the
kernel Makefile to understand version/patchlevel. If this
should fail it's a minor issue, it's just to support kernels
< 2.5; there shouldn't be many of those out there.