remotes/sonyir: Rename, split and revise key symbols.
The vendor name is sony, not sonyir. Renaming to conform with
the others.

Some of these programmable devices are quite complex, and having
more than one in a single config file is perhaps not a good idea.

Revising symbols using lirc-config-tool, renaming keys starting
with KEY_ to X_KEY_.... The basic contract is that a symbol
starting with KEY_ is an official symbol.